Clean up ChildForm and FormField so ChildForm is the primary element and FormField inherits it.
Also simplify ChildForm input data so it aligns with the data expected in the ManyForms list
